+++ libs/common/src/lib/interfaces/admin-market-data-details.interface.ts
@@ -3,6 +3,6 @@ import { MarketData } from '@prisma/client';
 import { EnhancedSymbolProfile } from './enhanced-symbol-profile.interface';
 
 export interface AdminMarketDataDetails {
-  assetProfile: EnhancedSymbolProfile;
+  assetProfile: Partial<EnhancedSymbolProfile>;
   marketData: MarketData[];
 }